backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Processing
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2  # Excel files
PyPDF2==3.0.1    # PDF parsing
pdfplumber==0.10.3  # Advanced PDF parsing

# AI/ML
openai==1.10.0
anthropic==0.8.1  # Claude API
langchain==0.1.5

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0
python-dotenv==1.0.0

# API Integrations
plaid-python==16.0.0  # Banking API
razorpay==1.4.1       # Payment gateway
requests==2.31.0
httpx==0.26.0

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
faker==22.6.0

# Monitoring & Logging
loguru==0.7.2
